cloud/requirements.txt
fastapi
uvicorn
numpy
scikit-learn
google-cloud-firestore
python-dotenv
requests
paho-mqtt

